--- src/main/kotlin/com/ogp404/ogps/reports_api/user/repository/AdminEntityRepository.kt
@@ -1,10 +1,0 @@
-package com.ogp404.ogps.reports_api.user.repository
-
-import com.ogp404.ogps.reports_api.user.repository.entity.AdminEntity
-import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.stereotype.Repository
-
-@Repository
-interface AdminEntityRepository : JpaRepository<AdminEntity, Int>{
-
-}
