+++ src/main/kotlin/com/ogp404/ogps/reports_api/user/repository/AdminRepository.kt
@@ -1,10 +1,10 @@
 package com.ogp404.ogps.reports_api.user.repository
 
-import com.ogp404.ogps.reports_api.user.repository.entity.UserEntity
+import com.ogp404.ogps.reports_api.user.repository.entity.Admin
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface UserEntityRepository : JpaRepository<UserEntity, Int>{
+interface AdminRepository : JpaRepository<Admin, Int>{
 
 }
