+++ src/main/kotlin/com/ogp404/ogps/reports_api/user/repository/UserRepository.kt
@@ -1,17 +1,10 @@
 package com.ogp404.ogps.reports_api.user.repository
 
-import com.ogp404.ogps.reports_api.user.repository.entity.Person
-import org.springframework.data.jpa.repository.Query
-import org.springframework.data.repository.CrudRepository
+import com.ogp404.ogps.reports_api.user.repository.entity.User
+import org.springframework.data.jpa.repository.JpaRepository
+import org.springframework.stereotype.Repository
 
-interface UserRepository : CrudRepository<Person, Int> {
+@Repository
+interface UserRepository : JpaRepository<User, Int>{
 
-    @Query(value = "SELECT * FROM person WHERE correo=?1", nativeQuery = true)
-    fun findByEmail(email: String): Person?
-
-    @Query(value = "SELECT * FROM person WHERE mail=?1 AND password=?2", nativeQuery = true)
-    fun findByEmailAndPassword(email: String, password: String): Person?
-
-    @Query(value = "SELECT * FROM person WHERE token=?1", nativeQuery = true)
-    fun findByToken(token: String): Person?
 }
